Update class documentation and move Subscription methods to appropriate class.
