+++ TwitchLib/TwitchAPI/Follows.cs
@@ -1,5 +1,8 @@
 ﻿namespace TwitchLib.TwitchAPI
 {
+    /// <summary>
+    ///     Status of follow relationships between users and channels.
+    /// </summary>
     public class Follows : ApiBase
     {
     }
