+++ TwitchLib/TwitchAPI/Games.cs
@@ -1,5 +1,9 @@
 ﻿namespace TwitchLib.TwitchAPI
 {
+    /// <summary>
+    ///     Games are categories (e.g. League of Legends, Diablo 3) used by streams and channels. Games can be searched for by
+    ///     query.
+    /// </summary>
     public class Games : ApiBase
     {
     }
