+++ TwitchLib/TwitchAPI/Teams.cs
@@ -5,11 +5,14 @@ using Newtonsoft.Json.Linq;
 
 namespace TwitchLib.TwitchAPI
 {
+    /// <summary>
+    ///     Teams are an organization of channels.
+    /// </summary>
     public class Teams : ApiBase
     {
         /// <summary>
-        /// Retrieves a TwitchTeamMember list of all members in a Twitch team.
-        /// <para>Note: This uses an undocumented API endpoint and reliability is not guaranteed.</para>
+        ///     Retrieves a TwitchTeamMember list of all members in a Twitch team.
+        ///     <para>Note: This uses an undocumented API endpoint and reliability is not guaranteed.</para>
         /// </summary>
         /// <param name="teamName">The name of the Twitch team to search for.</param>
         /// <returns>A TwitchTeamMember list of all members in a Twitch team.</returns>
