+++ TwitchLib/TwitchAPI/Users.cs
@@ -1,16 +1,22 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using System.Net;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
 
 namespace TwitchLib.TwitchAPI
 {
+    /// <summary>
+    ///     These are members of the Twitch community who have a Twitch account. If broadcasting, they can own a stream that
+    ///     they can broadcast on their channel. If mainly viewing, they might follow or subscribe to channels.
+    /// </summary>
     public class Users : ApiBase
     {
         /// <summary>
-        /// Retrieves a string list of channels hosting a specified channel.
-        /// <para>Note: This uses an undocumented API endpoint and reliability is not guaranteed. Additionally, this makes 2 API calls so limited use is recommended.</para>
+        ///     Retrieves a string list of channels hosting a specified channel.
+        ///     <para>
+        ///         Note: This uses an undocumented API endpoint and reliability is not guaranteed. Additionally, this makes 2
+        ///         API calls so limited use is recommended.
+        ///     </para>
         /// </summary>
         /// <param name="channel">The name of the channel to search for.</param>
         /// <returns>A list of all channels that are currently hosting the specified channel.</returns>
@@ -27,7 +33,7 @@ namespace TwitchLib.TwitchAPI
         }
 
         /// <summary>
-        /// Retrieves whether a specified user is following the specified user.
+        ///     Retrieves whether a specified user is following the specified user.
         /// </summary>
         /// <param name="username">The user to check the follow status of.</param>
         /// <param name="channel">The channel to check against.</param>
