+++ TwitchLib/TwitchAPI/Videos.cs
@@ -1,5 +1,9 @@
 ﻿namespace TwitchLib.TwitchAPI
 {
+    /// <summary>
+    ///     Videos are broadcasts or chapters owned by a channel. Broadcasts are unedited videos that are saved after a
+    ///     streaming session. Chapters are videos edited from broadcasts by the channel's owner.
+    /// </summary>
     public class Videos : ApiBase
     {
     }
